feat: add logger service and profile service with CRUD operations
- Implemented logger service using winston for structured logging.
- Added request logging middleware to capture request details.
- Created profile service with functions for fetching, creating, and managing profiles.
- Integrated external API calls for gender, age, and nationality data.
- Added CSV upload and export functionality for profiles.
- Introduced error handling and validation for profile creation and CSV processing.
- Updated package.json for winston and readable-stream dependencies.

+++ scripts/migrate.js
@@ -18,7 +18,7 @@ const migrations = [
     sql: `
       CREATE TABLE IF NOT EXISTS users (
         id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
-        github_id INTEGER UNIQUE NOT NULL,
+        github_id BIGINT UNIQUE NOT NULL,
         username VARCHAR(255) UNIQUE NOT NULL,
         email VARCHAR(255),
         avatar_url VARCHAR(500),
@@ -28,7 +28,7 @@ const migrations = [
         created_at TIMESTAMP DEFAULT NOW(),
         updated_at TIMESTAMP DEFAULT NOW()
       );
-      
+
       CREATE INDEX IF NOT EXISTS idx_users_github_id ON users(github_id);
       CREATE INDEX IF NOT EXISTS idx_users_username ON users(username);
     `
